fixed a bug in report and a bug in project deletion confirmation

+++ PTTMobile4/app/src/main/java/com/example/pttmobile4/models/Report.java
@@ -16,7 +16,7 @@ public class Report {
     private int completedPomodoros;
     @SerializedName("totalHoursWorkedOnProject")
     @Expose
-    private int totalHoursWorkedOnProject;
+    private double totalHoursWorkedOnProject;
 
 
     public ArrayList<Report_sessions> getSessions() {
@@ -27,7 +27,7 @@ public class Report {
         return completedPomodoros;
     }
 
-    public int getTotalHoursWorkedOnProject() {
+    public double getTotalHoursWorkedOnProject() {
         return totalHoursWorkedOnProject;
     }
 }
